refactor(python): Move `map_dict` implementation
