+++ py-polars/polars/expr/operations/__init__.py
@@ -1,0 +1,5 @@
+from polars.expr.operations.map_dict import map_dict
+
+__all__ = [
+    "map_dict",
+]
